requirements.txt
paho-mqtt==2.1.0
Flask==3.1.0
python-socketio==5.12.1
Flask-SocketIO==5.5.1
openai==1.63.2

mijiaAPI==1.3.12